feat: migrate provider to upjet no-fork Terraform Plugin SDK architecture
Wire config/provider.go to run resources through upjet's TerraformPluginSDK
no-fork mode (WithTerraformProvider + WithTerraformPluginSDKIncludeList)
instead of the CLI-fork Terraform runtime; empty the CLI-fork include list.
Add a shared OperationTrackerStore across cluster/namespaced controllers in
cmd/provider/main.go.

Bump vendored Snowflake TF provider 2.18.0 -> 2.19.0. Replace the Makefile's
pull-docs target with fetch-snowflake-provider-src, which vendors the
provider source once (patched for upstream's v2 go.mod module-path bug) and
serves both doc scraping and Go library import, instead of cloning twice.
Bump go.mod to 1.26.4 and pull in terraform-plugin-sdk/v2 as a direct dep.

Regenerate zz_* API types/controllers and package/crds/*.yaml against the
new provider schema (e.g. warehouse numeric fields number -> integer/int64,
simplified field descriptions).

+++ apis/cluster/preview/v1alpha1/zz_alert_types.go
@@ -32,7 +32,7 @@ type AlertInitParameters struct {
 	Condition *string `json:"condition,omitempty" tf:"condition,omitempty"`
 
 	// (Boolean) (Default: false) Specifies if an alert should be 'started' (enabled) after creation or should remain 'suspended' (default).
-	// (Default: `false`) Specifies if an alert should be 'started' (enabled) after creation or should remain 'suspended' (default).
+	// Specifies if an alert should be 'started' (enabled) after creation or should remain 'suspended' (default).
 	Enabled *bool `json:"enabled,omitempty" tf:"enabled,omitempty"`
 
 	// (String) The warehouse the alert will use.
@@ -63,7 +63,7 @@ type AlertObservation struct {
 	Database *string `json:"database,omitempty" tf:"database,omitempty"`
 
 	// (Boolean) (Default: false) Specifies if an alert should be 'started' (enabled) after creation or should remain 'suspended' (default).
-	// (Default: `false`) Specifies if an alert should be 'started' (enabled) after creation or should remain 'suspended' (default).
+	// Specifies if an alert should be 'started' (enabled) after creation or should remain 'suspended' (default).
 	Enabled *bool `json:"enabled,omitempty" tf:"enabled,omitempty"`
 
 	// (String) Fully qualified name of the resource. For more information, see object name resolution.
@@ -110,7 +110,7 @@ type AlertParameters struct {
 	Database *string `json:"database" tf:"database,omitempty"`
 
 	// (Boolean) (Default: false) Specifies if an alert should be 'started' (enabled) after creation or should remain 'suspended' (default).
-	// (Default: `false`) Specifies if an alert should be 'started' (enabled) after creation or should remain 'suspended' (default).
+	// Specifies if an alert should be 'started' (enabled) after creation or should remain 'suspended' (default).
 	// +kubebuilder:validation:Optional
 	Enabled *bool `json:"enabled,omitempty" tf:"enabled,omitempty"`
 
@@ -133,7 +133,7 @@ type AlertScheduleInitParameters struct {
 
 	// (Number) Specifies the interval in minutes for the alert schedule. The interval must be greater than 0 and less than 1440 (24 hours).
 	// Specifies the interval in minutes for the alert schedule. The interval must be greater than 0 and less than 1440 (24 hours).
-	Interval *float64 `json:"interval,omitempty" tf:"interval,omitempty"`
+	Interval *int64 `json:"interval,omitempty" tf:"interval,omitempty"`
 }
 
 type AlertScheduleObservation struct {
@@ -144,7 +144,7 @@ type AlertScheduleObservation struct {
 
 	// (Number) Specifies the interval in minutes for the alert schedule. The interval must be greater than 0 and less than 1440 (24 hours).
 	// Specifies the interval in minutes for the alert schedule. The interval must be greater than 0 and less than 1440 (24 hours).
-	Interval *float64 `json:"interval,omitempty" tf:"interval,omitempty"`
+	Interval *int64 `json:"interval,omitempty" tf:"interval,omitempty"`
 }
 
 type AlertScheduleParameters struct {
@@ -157,7 +157,7 @@ type AlertScheduleParameters struct {
 	// (Number) Specifies the interval in minutes for the alert schedule. The interval must be greater than 0 and less than 1440 (24 hours).
 	// Specifies the interval in minutes for the alert schedule. The interval must be greater than 0 and less than 1440 (24 hours).
 	// +kubebuilder:validation:Optional
-	Interval *float64 `json:"interval,omitempty" tf:"interval,omitempty"`
+	Interval *int64 `json:"interval,omitempty" tf:"interval,omitempty"`
 }
 
 type CronInitParameters struct {
